# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NEXA API",
    "version": "0.1.0",
    "description": "NEXA Workplace - internal social + communication platform. All /api/v1 endpoints except auth require `Authorization: Bearer <access_token>`."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "status": {
            "type": "integer",
            "example": 404
          },
          "error": {
            "type": "string",
            "example": "Post not found"
          },
          "code": {
            "type": "string",
            "example": "POST_NOT_FOUND"
          },
          "details": {
            "nullable": true
          },
          "request_id": {
            "type": "string",
            "example": "req_0f8fad5b-d9cb-469f-a165-70867728950e"
          }
        },
        "required": [
          "success",
          "status",
          "error",
          "code",
          "request_id"
        ]
      },
      "Me": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserProfile"
          },
          {
            "type": "object",
            "properties": {
              "last_login_at": {
                "type": "string",
                "nullable": true,
                "format": "date-time"
              }
            },
            "required": [
              "last_login_at"
            ]
          }
        ]
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DEACTIVATED"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "display_name",
          "avatar_url",
          "bio",
          "status",
          "created_at",
          "updated_at"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "pattern": "^[a-zA-Z0-9._-]+$",
            "example": "an.nguyen"
          },
          "email": {
            "type": "string",
            "example": "an.nguyen@nexa.io"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "example": "secret123"
          },
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "An Nguyễn"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "LoginResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AuthTokens"
          },
          {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/Me"
              }
            },
            "required": [
              "user"
            ]
          }
        ]
      },
      "AuthTokens": {
        "type": "object",
        "properties": {
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "description": "Access token lifetime in seconds"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": [
          "token_type",
          "access_token",
          "expires_in",
          "refresh_token"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "an.nguyen@nexa.io"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ExternalLoginResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/LoginResult"
          },
          {
            "type": "object",
            "properties": {
              "created": {
                "type": "boolean",
                "description": "Whether this sign-in created the account"
              }
            },
            "required": [
              "created"
            ]
          }
        ]
      },
      "GoogleSignInRequest": {
        "type": "object",
        "properties": {
          "id_token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8192,
            "description": "The ID token Google returned for this sign-in"
          },
          "nonce": {
            "type": "string",
            "minLength": 16,
            "maxLength": 256,
            "description": "The nonce sent in the authorization request; the token must carry it"
          }
        },
        "required": [
          "id_token",
          "nonce"
        ]
      },
      "LinkAccountRequest": {
        "type": "object",
        "properties": {
          "link_token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096,
            "description": "From the ACCOUNT_LINK_REQUIRED error; valid for 10 minutes"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          }
        },
        "required": [
          "link_token",
          "password"
        ]
      },
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "current_password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          },
          "new_password": {
            "type": "string",
            "minLength": 6,
            "example": "secret123"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ]
      },
...
    },
...
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new account",
        "description": "The new user joins the default organization; its first member becomes OWNER. Also served under `/api` without the `/v1` segment (exam contract, ADR-010).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Me"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log in with email and password",
        "description": "Also served under `/api` without the `/v1` segment (exam contract, ADR-010).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens and profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginResult"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/oauth/google": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign in with a Google ID token",
        "description": "The token must be issued for this server's GOOGLE_CLIENT_ID and carry the nonce of the authorization request. A new email gets an account (joining the default organization). If the email already has an account, answers 409 ACCOUNT_LINK_REQUIRED with a `link_token`: confirm that account's password with /auth/oauth/link to connect Google. 404 SSO_NOT_CONFIGURED when Google sign-in is off.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSignInRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens and profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ExternalLoginResult"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/oauth/link": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Connect Google to an existing account by confirming its password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens and profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginResult"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange a refresh token for a new token pair",
        "description": "The refresh token is rotated: the old one stops working. Replaying an already-used token revokes the whole session.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthTokens"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "End the session a refresh token belongs to",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Logged out (idempotent)"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password and sign out all other sessions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}